Move classes into respective packages
