+++ data/duke.txt
@@ -3,4 +3,3 @@ deadline ### false ### dashboard ### 2023-01-25
 event ### false ### climbing camp ### 2023-01-26 ### 2023-01-28
 todo ### true ### cs2103 Level-10
 deadline ### false ### passport renewal ### 2023-01-27
-todo ### true ### find home
