+++ src/main/java/duke/exceptions/InvalidIndexException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidIndexException extends DukeException {
     public InvalidIndexException(int idx) {
         super(String.format("Index of %d is out of bounds!", idx));
