+++ src/main/java/duke/exceptions/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class UnknownCommandException extends DukeException {
     private String command;
