+++ src/main/java/duke/task/Deadline.java
@@ -1,6 +1,9 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+import duke.exceptions.InvalidDateException;
 
 public class Deadline extends Task {
     private LocalDate deadline;
