+++ src/main/java/duke/task/Event.java
@@ -1,6 +1,9 @@
+package duke.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+import duke.exceptions.InvalidDateException;
 
 public class Event extends Task {
     private LocalDate startDate;
