+++ src/main/java/duke/ui/Span.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 public class Span {
     public static String format(String s) {
         String[] sp = s.split("\n");
